packages/python/nemesiscommon/pyproject.toml
[tool.poetry]
name = "nemesiscommon"
version = "0.2.2"
description = "Common nemesis helper functions"
authors = ["SpecterOps <specterops.io>"]

[tool.poetry.dependencies]
python = "^3.9"
nemesispb = { path = "../nemesispb", develop = true }
boto3 = "^1.24.59"
pycryptodomex = "^3.15.0"
requests = "^2.28.1"
pydantic = "^1.10.4"
tenacity = "^8.0.1"
aio-pika = "^8.3.0"
pyyaml = "^6.0"
typing-extensions = "^4.4.0"
aioboto3 = "^10.4.0"
structlog = "^23.1.0"
rich = "^13.3.3"
yara-python = "^4.3.0"
prometheus-async = {extras = ["aiohttp"], version = "^22.2.0"}
winacl = "^0.1.7"
miniopy-async = "^1.15"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
flake8 = "^5.0.4"
mypy = "^0.971"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
import_heading_stdlib = "Standard Libraries"
import_heading_firstparty = "Nemesis Libraries"
import_heading_thirdparty = "3rd Party Libraries"
profile = "black"
extra_standard_library = "asgiref"
known_first_party = "nemesis"
src_paths = ["isort", "test"]
line_length = 90
use_parentheses = true
multi_line_output = 3
include_trailing_comma = true
ensure_newline_before_comments = true
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

# [tool.pyright]
# Using pyrightconfig.json instead since pyproject.toml has 2nd class
# support in VS code: https://github.com/microsoft/pyright/issues/694#issuecomment-823778235

[tool.black]
line-length = 240
